third_party/blink/web_tests/plugins/drag-events-expected.txt

CONSOLE MESSAGE: Dragging over plugin_drag
Plugin received event: MouseEnter
Plugin received event: MouseMove
Plugin received event: DragEnter
Plugin received event: DragOver
Plugin received event: DragLeave
Plugin received event: DragEnter
Plugin received event: DragOver
Plugin received event: DragDrop
Plugin received event: MouseLeave
CONSOLE MESSAGE: Dragging over plugin_editable
Plugin received event: MouseEnter
Plugin received event: MouseMove
Plugin received event: MouseLeave
CONSOLE MESSAGE: Dragging over plugin
Plugin received event: MouseEnter
Plugin received event: MouseMove
Plugin received event: MouseLeave
  
Some text to drag
